requirements.txt
## 11. **`requirements.txt`** (Dependencies)

google-ads
rich
PyQt5
schedule
pandas
tensorflow
scikit-learn
numpy
